Add app-covid-visu Dashboard Covid cases & forecast for France and South Korea
This application is a dashboard in Flask Python to follow evolution of COVID cases in France and South Korea. 
A deep learning model is able to estimate the evolution for next 2 weeks. 
If the estimation is under reality, it means that the situation is worst than before...
My github open source project is here : https://github.com/jeugregg/coronavirusModel